requirements.txt
lighter-sdk>=1.0.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
numpy>=1.26.0
pandas>=2.1.0
